+++ Sources/DKGame/Graphics/ShaderBindingSet.swift
@@ -6,14 +6,14 @@
 //
 
 public struct ShaderBinding {
-    public var binding: UInt32
+    public var binding: Int
     public var type: ShaderDescriptorType
-    public var arrayLength: UInt32
+    public var arrayLength: Int
     public var immutableSamplers: SamplerState?
 
-    public init(binding: UInt32,
+    public init(binding: Int,
                 type: ShaderDescriptorType,
-                arrayLength: UInt32,
+                arrayLength: Int,
                 immutableSamplers: SamplerState? = nil) {
         self.binding = binding
         self.type = type
@@ -33,12 +33,12 @@ public struct ShaderBindingSetLayout {
 
 public struct GPUBufferBindingInfo {
     public var buffer: Buffer
-    public var offset: UInt64
-    public var length: UInt64
+    public var offset: Int
+    public var length: Int
 
     public init(buffer: Buffer,
-                offset: UInt64,
-                length: UInt64) {
+                offset: Int,
+                length: Int) {
         self.buffer = buffer
         self.offset = offset
         self.length = length
@@ -47,16 +47,16 @@ public struct GPUBufferBindingInfo {
 
 public protocol ShaderBindingSet {
     // bind buffers
-    func setBuffer(_: Buffer, offset: UInt64, length: UInt64, binding: UInt32)
-    func setBufferArray(_ : [GPUBufferBindingInfo], binding: UInt32)
+    func setBuffer(_: Buffer, offset: Int, length: Int, binding: Int)
+    func setBufferArray(_ : [GPUBufferBindingInfo], binding: Int)
 
     // bind textures
-    func setTexture(_: Texture, binding: UInt32)
-    func setTextureArray(_: [Texture], binding: UInt32)
+    func setTexture(_: Texture, binding: Int)
+    func setTextureArray(_: [Texture], binding: Int)
 
     // bind samplers
-    func setSamplerState(_: SamplerState, binding: UInt32)
-    func setSamplerStateArray(_: [SamplerState], binding: UInt32)
+    func setSamplerState(_: SamplerState, binding: Int)
+    func setSamplerStateArray(_: [SamplerState], binding: Int)
 
     var device: GraphicsDevice { get }
 }
